requirements.txt
cohere==4.26.1
firebase_admin==6.2.0
fitz==0.0.1.dev2
icecream==2.1.3
python-dotenv==1.0.0
Requests==2.31.0
